add Difficulties container and Difficulty,TextInput components
