+++ src/components/TextInput/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function TextInput({ type = "text", ...props }) {
+  return <input type={type} {...props} />;
+}
+
+export default TextInput;
